feat: cloud logs
